refactor: rename types to base
