+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './ArcxAnalyticsProvider'
 export * from './ArcxAnalyticsSdk'
-export * from './types/types'
+export * from './types'
